backend/requirements.txt
asgiref==3.8.1
Django==4.2.15
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
